Fixing test

+++ packages/tegh-server/src/core/spool/sagas/taskDeletionSaga.js
@@ -6,12 +6,15 @@ import getTasksPendingDeletion from '../selectors/getTasksPendingDeletion'
 const taskDeletionSaga = function*() {
   const { takeLatest, select, put } = effects
 
-  yield takeLatest('*', function*() {
+  yield takeLatest('*', function*(action) {
+    if (action.type === DELETE_TASKS) return
+
     /* delete all the tasks pending deletion */
     const tasks = yield select(getTasksPendingDeletion)
 
     if (tasks.size === 0) return
 
+    // throw new Error('TASKS EXIST' + JSON.stringify(deleteTasks({ ids: tasks.map(task => task.id) })))
     yield put(deleteTasks({ ids: tasks.map(task => task.id) }))
   })
 }
